Move tests touching the global environment into integration tests
