+++ tests/PHPUnit/Unit/CronArchive/FixedSiteIdsTest.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Tests\Integration\CronArchive;
+namespace Piwik\Tests\Unit\CronArchive;
 
 use Piwik\CronArchive\FixedSiteIds;
